WaE 0xdeadbeaf 32 bits litteral casting to 64 bits pointer

+++ cppu/source/uno/destr.hxx
@@ -132,7 +132,7 @@ inline void _destructAny(
         break;
     }
 #if OSL_DEBUG_LEVEL > 0
-    pAny->pData = reinterpret_cast<void *>(0xdeadbeef);
+    pAny->pData = reinterpret_cast<void *>((uintptr_t)0xdeadbeef);
 #endif
 
     ::typelib_typedescriptionreference_release( pType );
